Move attribute group services into proper namespace
