+++ src/PrestaShopBundle/Resources/config/services/adapter/attribute_group.yml
@@ -2,14 +2,20 @@ services:
     _defaults:
         public: true
 
-    prestashop.adapter.attribute_group.command_handler.delete_attribute_group_handler:
-        class: 'PrestaShop\PrestaShop\Adapter\AttributeGroup\CommandHandler\DeleteAttributeGroupHandler'
+    prestashop.adapter.product.attribute_group.attribute_group_view_data_provider:
+        class: 'PrestaShop\PrestaShop\Adapter\Product\AttributeGroup\AttributeGroupViewDataProvider'
+        arguments:
+            - "@=service('prestashop.adapter.legacy.context').getContext().language.id"
+            - "@=service('prestashop.adapter.legacy.configuration')"
+
+    prestashop.adapter.product.attribute_group.command_handler.delete_attribute_group_handler:
+        class: 'PrestaShop\PrestaShop\Adapter\Product\AttributeGroup\CommandHandler\DeleteAttributeGroupHandler'
         tags:
             - name: 'tactician.handler'
               command: 'PrestaShop\PrestaShop\Core\Domain\Product\AttributeGroup\Command\DeleteAttributeGroupCommand'
 
-    prestashop.adapter.attribute_group.command_handler.bulk_delete_attribute_group_handler:
-        class: 'PrestaShop\PrestaShop\Adapter\AttributeGroup\CommandHandler\BulkDeleteAttributeGroupHandler'
+    prestashop.adapter.product.attribute_group.command_handler.bulk_delete_attribute_group_handler:
+        class: 'PrestaShop\PrestaShop\Adapter\Product\AttributeGroup\CommandHandler\BulkDeleteAttributeGroupHandler'
         tags:
             - name: 'tactician.handler'
               command: 'PrestaShop\PrestaShop\Core\Domain\Product\AttributeGroup\Command\BulkDeleteAttributeGroupCommand'
