--- src/PrestaShopBundle/Resources/config/services/core/attribute_group.yml
@@ -1,9 +1,0 @@
-services:
-  _defaults:
-    public: true
-
-  prestashop.core.attribute_group.attribute_group_view_data_provider:
-    class: 'PrestaShop\PrestaShop\Adapter\AttributeGroup\AttributeGroupViewDataProvider'
-    arguments:
-      - "@=service('prestashop.adapter.legacy.context').getContext().language.id"
-      - "@=service('prestashop.adapter.legacy.configuration')"
